update funtion DanhSachCanMua

+++ GoceryStore_DACN/DTOs/ThucPhamDTO.cs
@@ -9,7 +9,7 @@ namespace GoceryStore_DACN.DTOs
         [Required]
         public int ID_LoaiThucPham { get; set; }
         public string DVT { get; set; } = "Kg";
-        public int SoLuong { get; set; } = 0;
+        public double SoLuong { get; set; } = 0;
         public double GiaBan { get; set; } = 0;
         public IFormFile ImageFile { get; set; } = null;
         public string TrangThai { get; set; } = "Còn Hàng";
